HSEARCH-3735 Expose the concept of workset to LuceneIndexIndexingPlan
