+++ backend/lucene/src/main/java/org/hibernate/search/backend/lucene/orchestration/impl/LuceneSingleWriteWorkSet.java
@@ -1,0 +1,40 @@
+/*
+ * Hibernate Search, full-text search for your domain model
+ *
+ * License: GNU Lesser General Public License (LGPL), version 2.1 or later
+ * See the lgpl.txt file in the root directory or <http://www.gnu.org/licenses/lgpl-2.1.html>.
+ */
+package org.hibernate.search.backend.lucene.orchestration.impl;
+
+import java.util.concurrent.CompletableFuture;
+
+import org.hibernate.search.backend.lucene.work.impl.LuceneWriteWork;
+import org.hibernate.search.engine.backend.work.execution.DocumentCommitStrategy;
+import org.hibernate.search.engine.backend.work.execution.DocumentRefreshStrategy;
+
+class LuceneSingleWriteWorkSet<T> implements LuceneWriteWorkSet {
+	private final LuceneWriteWork<T> work;
+	private final CompletableFuture<T> future;
+	private final DocumentCommitStrategy commitStrategy;
+	private final DocumentRefreshStrategy refreshStrategy;
+
+	LuceneSingleWriteWorkSet(LuceneWriteWork<T> work, CompletableFuture<T> future,
+			DocumentCommitStrategy commitStrategy, DocumentRefreshStrategy refreshStrategy) {
+		this.work = work;
+		this.future = future;
+		this.commitStrategy = commitStrategy;
+		this.refreshStrategy = refreshStrategy;
+	}
+
+	@Override
+	public void submitTo(LuceneWriteWorkProcessor processor) {
+		processor.beforeWorkSet( commitStrategy, refreshStrategy );
+		T result = processor.submit( work );
+		processor.afterWorkSet( future, result );
+	}
+
+	@Override
+	public void markAsFailed(Throwable t) {
+		future.completeExceptionally( t );
+	}
+}
