+++ backend/lucene/src/main/java/org/hibernate/search/backend/lucene/orchestration/impl/LuceneWriteWorkSet.java
@@ -1,0 +1,12 @@
+/*
+ * Hibernate Search, full-text search for your domain model
+ *
+ * License: GNU Lesser General Public License (LGPL), version 2.1 or later
+ * See the lgpl.txt file in the root directory or <http://www.gnu.org/licenses/lgpl-2.1.html>.
+ */
+package org.hibernate.search.backend.lucene.orchestration.impl;
+
+import org.hibernate.search.engine.backend.orchestration.spi.BatchingExecutor;
+
+public interface LuceneWriteWorkSet extends BatchingExecutor.WorkSet<LuceneWriteWorkProcessor> {
+}
